Refactor SharpenPipe and Grayscale processing, update dependencies
Reworked the SharpenPipe to be more efficient and modular. Updated Grayscale processing for enhanced performance. Cleaned up and reorganized imports for better readability. Also updated necessary dependencies in the processing crates.

+++ crates/core/src/threading/progress.rs
@@ -1,21 +1,21 @@
-use std::sync::atomic::{AtomicU32, Ordering};
+use std::sync::atomic::{AtomicUsize, Ordering};
 
 pub struct Progress {
-    value: AtomicU32,
-    max_value: u32,
+    value: AtomicUsize,
+    max_value: usize,
     on_increment: Option<Box<dyn Fn() + Send + Sync>>,
 }
 
 impl Progress {
     pub fn new() -> Self {
         Progress {
-            value: AtomicU32::new(0),
+            value: AtomicUsize::new(0),
             max_value: 0,
             on_increment: None,
         }
     }
 
-    pub fn setup(&mut self, max_value: u32) {
+    pub fn setup(&mut self, max_value: usize) {
         self.max_value = max_value;
         self.value.store(0, Ordering::SeqCst);
     }
@@ -27,17 +27,17 @@ impl Progress {
         }
     }
 
-    pub fn get(&self) -> u32 {
+    pub fn get(&self) -> usize {
         self.value.load(Ordering::SeqCst)
     }
 
-    pub fn get_max(&self) -> u32 {
+    pub fn get_max(&self) -> usize {
         self.max_value
     }
     pub fn get_percentage(&self) -> f32 {
         (self.get() as f32 / self.get_max() as f32) * 100.0
     }
-    
+
     pub fn set_on_increment<F: Fn() + Send + Sync + 'static>(&mut self, on_increment: F) {
         self.on_increment = Some(Box::new(on_increment));
     }
